fix:extracted color pallete in pallate comp
