+++ src/components/index.js
@@ -5,3 +5,4 @@ export * from "./aside/Aside"
 export * from "./modal/CreateModal"
 export * from "./eachNote/EachNote"
 export * from "./modal/EditModal"
+export * from "./pallete/ColorPallete"
